refactoring getAllowedLanguages function
